Ajout de la possibilité d'afficher un nuage de points dans la vue 3D. Possibilité de donner une valeur à un uniform sans que celui ci ne possède de shaders, la mise à jour dans le shader se fait à la création de celui-ci. Modification de la correction des NaNs dans VoxelAnalysis, ajout d'une limite de passes et modification des conditions (à tester)

+++ trunk/AMAPVox/Voxelisation/src/main/java/fr/amap/amapvox/voxelisation/configuration/VoxelParameters.java
@@ -59,7 +59,7 @@ public class VoxelParameters {
     private int rasterBandNumber;
     
     private boolean correctNaNsMode2;
-    private int correctNaNsNbSamplingThreshold;
+    private float correctNaNsNbSamplingThreshold;
     
     private int ladEstimationMode;
     private Type ladType;
@@ -309,11 +309,11 @@ public class VoxelParameters {
         this.correctNaNsMode2 = correctNaNsMode2;
     }
 
-    public int getCorrectNaNsNbSamplingThreshold() {
+    public float getCorrectNaNsNbSamplingThreshold() {
         return correctNaNsNbSamplingThreshold;
     }
 
-    public void setCorrectNaNsNbSamplingThreshold(int correctNaNsNbSamplingThreshold) {
+    public void setCorrectNaNsNbSamplingThreshold(float correctNaNsNbSamplingThreshold) {
         this.correctNaNsNbSamplingThreshold = correctNaNsNbSamplingThreshold;
     }
